Added file checks
Added checks to see if the file existed, if it does, then we use that file. If it doesnt exist, then we ask the user to select one. Once they select a file we will set the value of the path control to the selected file and we will use the newly selected value.