Changed when tip is set down on surface
Changed when tip is set down on surface from the lithography object points to the initialize lithography object. Also put an "Application Out" for the Lithography Objects to make sure the same wire is passed around regarding the application reference. Also removed the setting tip voltage to initial value as that was a carry over from programming the NSES2, and not needed for the asylum.