Update AFM_Lithography.vi
Minor change that will make the imaq image zoom to fit on start up, as somehow it was zoomed in everytime we started it up. Now we have it programmatically set to zoom to fit upon start up. This is more for developers, as hopefully the end users will be using the EXE.