Added a DDS for debugging
Added a disable diagram structure to prevent the program from deleting the temporary files so we can look at them for debugging purposes. The structure is currently set so that it is deleting the temporary files, but I left it there for future just in case we need to use it again.