Multiple changes to make the software compatible with MFP-3D
I made changes for the software to handle exponents in the transformation tokenizer

Added a safety check for filling in paths that have zero as a value for the path fill spacing

Removed the part of the code that relies on the fact that filepaths in SVGs are displayed as URIs. Instead, it just treats them as filepaths.